Do not use atoi (#27)
* Squash and combine all 5 commits into 1

Replaced the calls to atoi to use strtol instead. A typo in an error handling message was fixed. Also removed the requirement to use a parameter to indicate the debug level for the -d option, instead users can repeat the letter d as part of the option to indicate the level. Updated the usage message to reflect this change too.

Co-authored-by: Brooks Davis <brooks@one-eyed-alien.net>

---------

Co-authored-by: Jessica Man <psjm3@cam.ac.uk>
Co-authored-by: Jessica Man <149842632+psjm3@users.noreply.github.com>
Co-authored-by: Brooks Davis <brooks@one-eyed-alien.net>