Rtld linkmap scan (#30)
Save compartment data of target process to sqlite database so that views can make queries to the db to get the data they need. The vm view (chericat -v) has been updated to add a new column to show the corresponding compart_id (-1 means no assigned compart_id). We are currently giving specific compart_ids to ld-elf (-2) and guards (-3) so that we can separate them on the visualisation graphs. The stacks are assigned the compart_ids obtained from rtld.

The python code has also been updated to add a new compartment overview graph.  To not cause confusion on the graphs we have removed the .bss as separate entries on the database (otherwise they will be shown as standalone nodes).